Chore[COMP345-78]: refactor item
