+++ ProjectSourceCode/Test/testitemcontainer.h
@@ -18,7 +18,7 @@
 #include <cppunit/CompilerOutputter.h>
 #include <cppunit/XmlOutputter.h>
 
-#include "..\itemcontainer.h"
+#include "..\Item\itemcontainer.h"
 
 // using namespace std;
 using namespace CppUnit;
